fix bug: UnboundLocalError when segment_list and row_list empty
There will occur the error when the segment_list and row_list are empty, then the total_vectors and total_ids are assignment before clarify. This happened for me when I create all of the partition tags for data, and give None for partition tags when I do the milvus to hdf5, cause the default partition tag None has no data, then the error will occur.